requirements.txt
darts>=0.25.0
matplotlib>=3.6.2
numpy>=1.22.3
pandas>=1.5.1
salesforce_merlion>=2.0.2
scikit_learn>=1.1.3
scipy>=1.8.0
statsmodels>=0.14.0
torch>=1.11.0
ray>=2.6.3
tqdm>=4.64.0
dash>=2.9.3
dash-bootstrap-components>=1.5.0
